Reorganise ir and add auto serialisation (#78)
* organise ir into separate files

* add serialization of ir nodes via to_dict

limitation: NIRNode that contains nested NIRNodes e.g. NIRGraph must
reimplement to_dict

* make ruff happy and sort imports

* add deserialization of ir nodes via from_dict and reflection

limitation: NIRNode that contains nested NIRNodes e.g. NIRGraph must
reimplement from_dict
be careful that h5py serialize List[Tuple[str]] into numpy array of
bytes

* merge read and write into serialization

* make sure from_dict is an inverse of to_dict

* fix Union and Dict type hint
